fix(firewall-config): fix rich rule creation
It's impossible to create a rich rule because the Rich_Rule constructor
is incorrecly invoked on lines 3404-3709 of src/firewall-config.in.
Moreover, Rich_Rule attributes are set for its instance, which, due to
the frozen state of Rich_Rule, leads to even more errors.

Basically, the fix is quite simple: just create a dictionary of
attributes and pass them to the Rich_Rule constuctor at once.

Signed-off-by: Oleksandr Zakharov <x.zakharov@proton.me>